pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rstatus"
version = "0.0.1"
authors = [
  { name="Pedro Agustin Vega", email="pedroagustinvega@proton.me" },
]
description = "Library to query Minecraft server data"
readme = "README.md"
requires-python = ">=3.7"
license = { file="LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PySocks>=1.7.1",
    "dnspython>=2.7.0",
]
